+++ src/lib/telemetry.ts
