+++ src/lib/worker/thickness.ts
