--- src/lib/worker/types/src/lib/worker/socketsLoader.d.ts
@@ -1,4 +1,0 @@
-import { type Solid } from 'replicad'
-import type { CuttleKey } from './config'
-import type Trsf from './modeling/transformation'
-export declare function keyHole(key: CuttleKey, trsf: Trsf): Promise<Solid>
