--- src/lib/worker/types/src/routes/beta/lib/thickness.d.ts
@@ -1,2 +1,0 @@
-import { Vector3 } from 'three'
-export declare function thickness(pta: Vector3, ptb: Vector3, ptc: Vector3, pba: Vector3, pbb: Vector3, pbc: Vector3): number
